:recycle: #35 ExternalLibrary 도메인에서 util, service 로 계층화 리팩토링
